requirements.txt
accelerate==1.7.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
Automat==25.4.16
beautifulsoup4==4.13.4
blis==1.3.0
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
confection==0.1.5
constantly==23.10.4
contourpy==1.3.2
cryptography==45.0.3
cssselect==1.3.0
cycler==0.12.1
cymem==2.0.11
datasets==3.6.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
dill==0.3.8
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.11.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.58.1
frozenlist==1.6.0
fsspec==2025.3.0
gradio==5.31.0
gradio_client==1.10.1
groovy==0.1.2
h11==0.16.0
hanlp==2.1.1
hanlp_common==0.0.23
hanlp_downloader==0.0.25
hanlp_trie==0.0.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inquirerpy==0.3.4
ipykernel==6.29.5
ipython==8.36.0
ipywidgets==8.1.6
itemadapter==0.11.0
itemloaders==1.3.2
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.14
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lightning-utilities==0.14.3
ltp==4.2.14
ltp-core==0.1.4
ltp-extension==0.1.13
lxml==5.4.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
murmurhash==1.0.13
nest-asyncio==1.6.0
networkx==3.3
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
orjson==3.10.18
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
parsel==1.10.0
parso==0.8.4
pexpect==4.9.0
pfzy==0.3.4
phrasetree==0.0.9
pillow==11.0.0
platformdirs==4.3.7
preshed==3.0.10
prompt_toolkit==3.0.51
propcache==0.3.1
Protego==0.4.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyDispatcher==2.0.7
pydub==0.25.1
Pygments==2.19.1
pynvml==12.0.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypinyin==0.54.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.5.1.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
queuelib==1.8.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rich==14.0.0
ruff==0.11.12
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
Scrapy==2.13.1
seaborn==0.13.2
selenium==4.33.0
semantic-version==2.10.0
sentence-transformers==4.1.0
sentencepiece==0.2.0
service-identity==24.2.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.1
termcolor==3.1.0
thinc==8.3.6
threadpoolctl==3.6.0
tldextract==5.3.0
tokenizers==0.21.1
tomli==2.2.1
tomlkit==0.13.2
toposort==1.5
torch==2.6.0+cu124
torchaudio==2.6.0+cu124
torchmetrics==0.11.4
torchvision==0.21.0+cu124
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.3
trio==0.30.0
trio-websocket==0.12.2
triton==3.2.0
Twisted==24.11.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
w3lib==2.3.1
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.0
zope.interface==7.2
